fb: move over public SDK headers to include/
