+++ glamor/glamor_addtraps.c
@@ -27,7 +27,7 @@
  */
 #include <dix-config.h>
 
-#include "fb/fbpict.h"
+#include "include/fbpict.h"
 
 #include "glamor_priv.h"
