+++ include/meson.build
@@ -432,6 +432,11 @@ if build_xorg
             'extension.h',
             'extinit.h',
             'extnsionst.h',
+            'fb.h',
+            'fboverlay.h',
+            'fbpict.h',
+            'fbrop.h',
+            'wfbrename.h',
             'fd_notify.h',
             'fourcc.h',
             'gc.h',
